+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     - ReferenceProvider
     - DocumentHighlightProvider
 - References of actions and symbol macros are seeked in the entire workspace.
+- Updated list of actions and effects.
 
 ### Fixed
 - ReferenceProvider and DocumentHighlightProvider now correctly highlight prefix of `____symbol_` variation.